Switched Purchased & Sold values over to sum from dbo.ParcelLedger entries of corresponding transaction type.

+++ Source/Rio.EFModels/Entities/LandownerUsageReport.cs
@@ -17,8 +17,8 @@ namespace Rio.EFModels.Entities
         public int AccountNumber { get; set; }
         public double? AcresManaged { get; set; }
         public decimal? Allocation { get; set; }
-        public int? Purchased { get; set; }
-        public int? Sold { get; set; }
+        public decimal? Purchased { get; set; }
+        public decimal? Sold { get; set; }
         public decimal? TotalSupply { get; set; }
         public decimal? UsageToDate { get; set; }
         public decimal? CurrentAvailable { get; set; }
